Move app logic to gameplay
