--- client/src/HumanPoseEstimation.css
@@ -1,28 +1,0 @@
-.app-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.webcam-container {
-  position: relative;
-  width: 640px;
-  height: 480px;
-  transform: scaleX(-1);
-}
-
-.capture-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.captured-image {
-  width: 600px;
-  height: 600px;
-  margin-bottom: 10px;
-  /* Add any other styles you want for the captured image container */
-}
