+++ client/src/ImagePoseEstimation.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useRef } from "react";
 import * as tf from "@tensorflow/tfjs";
 import * as posenet from "@tensorflow-models/posenet";
 import { drawKeypoints, drawSkeleton } from "./utilities";
+import "./App.css";
 
 const ImagePoseEstimation = ({ imageSrc, onImagePoseDetected }) => {
   const canvasRef = useRef(null);
